requirements.txt
streamlit==1.28.1
crewai==0.28.8
langchain==0.1.5
langchain-openai==0.0.8
langchain-community==0.0.20
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
google-search-results==2.4.2
openai==1.12.0
chromadb==0.4.22
numpy==1.24.3
pandas==2.0.3
pydantic==2.5.3
tiktoken==0.5.2

